Scale up differed
I changed the scale up:
I now shifts the graph to the x and y axis so at least on vertex has a x value of 0 and at least a vertex has a y value of 0.
Furthermore I added a button to scale the graph to a box size, this is atm a constant value set by the mainframe.